Steam kan nu inloggen. Zijn nu 2 modules aan toegevoegd om het dashboard te maken en de module getsteamuserinfo.py om de gebruikersnaam url en status van de user op te vragen
