+++ dashboard.py
@@ -1,0 +1,3 @@
+def dashboardwindow(name, avatarurl, status):
+    print("dashboard...")
+    print(f"dev: {name}, {avatarurl}, {status}")
