+++ main.py
@@ -1,28 +1,333 @@
-import requests
 
-def get_player_summaries(steam_ids):
-    url = f'https://api.steampowered.com/ISteamUser/GetPlayerSummaries/v0002/?key=B5A67039860C1613632C4795B6C36245&steamids={",".join(steam_ids)}'
-    response = requests.get(url)
-    return response.json()
+# This file was generated by the Tkinter Designer by Parth Jadhav
+# https://github.com/ParthJadhav/Tkinter-Designer
 
-if __name__ == "__main__":
-    # Replace 'STEAM_ID' with the actual Steam ID you want to retrieve information for
-    steam_ids = ['76561199478010920']
+from sys import platform
 
-    player_data = get_player_summaries(steam_ids)
+if platform == "linux" or platform == "linux2":
+    pass
+elif platform == "darwin":
+    pass
+elif platform == "win32":
+    import ctypes
+    ctypes.windll.shcore.SetProcessDpiAwareness(True)
 
-    if 'response' in player_data and 'players' in player_data['response']:
-        players_info = player_data['response']['players']
 
-        if players_info:  # Check if there are players in the list
-            for player in players_info:
-                print(f"Steam ID: {player['steamid']}")
-                print(f"Name: {player['personaname']}")
-                print(f"Profile URL: {player['profileurl']}")
-                print(f"Avatar: {player['avatarfull']}")
-                print(f"Status: {player['personastate']}")
-                print()
-        else:
-            print("No player found with the provided Steam ID.")
+
+
+from getsteamuserinfo import getsteamuserinfo
+
+
+
+# from tkinter import *
+# Explicit imports to satisfy Flake8
+from tkinter import Tk, Canvas, Entry, Text, Button, PhotoImage
+import tkinter as tk
+from PIL import Image, ImageTk
+import webbrowser
+
+
+from pathlib import Path
+OUTPUT_PATH = Path(__file__).parent
+ASSETS_PATH = OUTPUT_PATH / Path(r"assets\frame0")
+#path naar de gecurvde vierkanten pngs
+
+
+def relative_to_assets(path: str) -> Path:
+    return ASSETS_PATH / Path(path)
+#define wat hoort bij het path
+
+def add_image_to_canvas(canvas, image_path, x, y, width, height):
+    # Load the image and resize it
+    original_image = Image.open(image_path)
+    resized_image = original_image.resize((width, height))
+
+    # Create a PhotoImage object from the resized image
+    tk_image = ImageTk.PhotoImage(resized_image)
+
+    # Create an image item at the specified coordinates
+    canvas.create_image(x, y, anchor=tk.NW, image=tk_image)
+    return tk_image
+
+def know_more_clicked(event):
+    instructions = (
+        "https://steamid.io/"
+        )
+    webbrowser.open_new_tab(instructions)
+
+
+def steamidinput():
+    steaminput = text_widget.get("1.0", "end-1c").strip()
+    if len(steaminput) != 17:
+        time = canvas.create_text(
+            162.0,
+            1013.0,
+            anchor="nw",
+            text="We konden uw SteamID niet vinden. Zorg ervoor dat u het SteamID64 heeft",
+            fill="#FFFFFF",
+            font=("Motiva Sans Bold", 40 * -1)
+        )
+        window.after(4000, lambda: canvas.delete(time))
+
     else:
-        print("Error fetching player data.")
+        steamid64 = steaminput
+        status1 = getsteamuserinfo(steamid64)
+        if status1 == 0:
+            time = canvas.create_text(
+                262.0,
+                1013.0,
+                anchor="nw",
+                text="Er is iets mis gegaan. Probeer het later opnieuw",
+                fill="#FFFFFF",
+                font=("Motiva Sans Bold", 40 * -1)
+                #staat nog niet gealigned op de x as
+            )
+            window.after(4000, lambda: canvas.delete(time))
+        elif status1 == 1:
+            time = canvas.create_text(
+                340.0,
+                1013.0,
+                anchor="nw",
+                text="Er is geen steam gebruiker gevonden met dat SteamID64",
+                fill="#FFFFFF",
+                font=("Motiva Sans Bold", 40 * -1)
+            )
+            window.after(4000, lambda: canvas.delete(time))
+        else:
+            window.destroy()
+        return steamid64
+
+
+
+window = Tk()
+
+window.geometry("1834x1117")
+window.configure(bg = "#FFFFFF")
+window.title("Inloggen met Steam...")
+
+
+canvas = Canvas(
+    window,
+    bg = "#FFFFFF",
+    height = 1117,
+    width = 1834,
+    bd = 0,
+    highlightthickness = 0,
+    relief = "ridge"
+)
+
+canvas.place(x = 0, y = 0)
+canvas.create_rectangle(
+    0.0,
+    0.0,
+    1834.0,
+    1117.0,
+    fill="#102431",
+    outline="")
+
+canvas.create_text(
+    100.0,
+    48.0,
+    anchor="nw",
+    text="Inloggen met",
+    fill="#FFFFFF",
+    font=("Motiva Sans ExtraBold", 64 * -1)
+)
+image_path = "assets\logo_steam.png"
+image = add_image_to_canvas(canvas, image_path, x=540, y=50, width=267, height=77)
+
+
+
+
+
+entry_image_1 = PhotoImage(
+    file=relative_to_assets("entry_1.png"))
+entry_bg_1 = canvas.create_image(
+    917.0,
+    585.0,
+    image=entry_image_1
+)
+canvas.create_text(
+    100.0,
+    139.0,
+    anchor="nw",
+    text="This site is not associated with Valve Corp.",
+    fill="#CACACA",
+    font=("Motiva Sans Regular", 24 * -1)
+)
+
+entry_image_2 = PhotoImage(
+    file=relative_to_assets("entry_2.png"))
+entry_bg_2 = canvas.create_image(
+    601.0,
+    381.0,
+    image=entry_image_2
+)
+
+text_widget = Text(
+    bd=0,
+    bg="#33333D",
+    fg="#FFFFFF",
+    highlightthickness=0,
+    wrap="word",  #zorg ervoor dat de woorden binnen de box blijven
+    padx=10,  # padding op de box zodat de text niet buiten de box komt
+    pady=20,  # padding op de box zodat de text niet butien de box komt
+    font=("Motiva Sans Bold", 18),  # pas het font aan naar rubik
+)
+text_widget.place(
+    x=206.0,
+    y=336.0,
+    width=790.0,
+    height=88.0
+)
+
+
+
+
+
+canvas.create_text(
+    196.0,
+    293.0,
+    anchor="nw",
+    text="INLOGGEN MET STEAMID-64",
+    fill="#1999FF",
+    font=("Motiva Sans Medium", 24 * -1)
+)
+
+canvas.create_text(
+    1164.0,
+    293.0,
+    anchor="nw",
+    text="INLOGGEN MET EEN QR-CODE",
+    fill="#1999FF",
+    font=("Motiva Sans Medium", 24 * -1)
+)
+
+button_image_1 = PhotoImage(
+    file=relative_to_assets("button_1.png"))
+button_1 = Button(
+    image=button_image_1,
+    borderwidth=0,
+    highlightthickness=0,
+    relief="flat"
+)
+button_1.bind('<Button-1>', know_more_clicked)
+button_1.place(
+    x=425.0,
+    y=799.0,
+    width=365.0,
+    height=78.0
+)
+
+button_image_2 = PhotoImage(
+    file=relative_to_assets("button_2.png"))
+button_2 = Button(
+    image=button_image_2,
+    borderwidth=0,
+    highlightthickness=0,
+    command=lambda: steamidinput(),
+    relief="flat"
+)
+button_2.place(
+    x=370.0,
+    y=605.0,
+    width=475.0,
+    height=78.0
+)
+
+
+entry_image_3 = PhotoImage(
+    file=relative_to_assets("entry_3.png"))
+entry_bg_3 = canvas.create_image(
+    1372.0,
+    543.0,
+    image=entry_image_3
+)
+
+
+image_path = r"assets\QRCODE.png"
+image1 = add_image_to_canvas(canvas, image_path, x=1181, y=354, width=381, height=378)
+
+entry_image_4 = PhotoImage(
+    file=relative_to_assets("entry_4.png"))
+entry_bg_4 = canvas.create_image(
+    383.5,
+    740.5,
+    image=entry_image_4
+)
+entry_4 = Entry(
+    bd=0,
+    bg="#1999FF",
+    fg="#000716",
+    highlightthickness=0
+)
+entry_4.place(
+    x=212.5,
+    y=736.0,
+    width=342.0,
+    height=7.0
+)
+
+entry_image_5 = PhotoImage(
+    file=relative_to_assets("entry_5.png"))
+entry_bg_5 = canvas.create_image(
+    835.0,
+    740.5,
+    image=entry_image_5
+)
+entry_5 = Entry(
+    bd=0,
+    bg="#1999FF",
+    fg="#000716",
+    highlightthickness=0
+)
+entry_5.place(
+    x=661.5,
+    y=736.0,
+    width=347.0,
+    height=7.0
+)
+
+canvas.create_text(
+    588.0,
+    726.0,
+    anchor="nw",
+    text="OF",
+    fill="#818181",
+    font=("Motiva Sans Medium", 24 * -1)
+)
+
+# canvas.create_text(
+#     322.0,
+#     1013.0,
+#     anchor="nw",
+#     text="We konden geen gebruiker vinden met dat SteamID",
+#     fill="#FFFFFF",
+#     font=("Motiva Sans Bold", 48 * -1)
+# )
+
+
+canvas.create_text(
+    1209.0,
+    770.0,
+    anchor="nw",
+    text="Gebruik de mobiele Steam-app \nom via een QR-code in te loggen",
+    fill="#FFFFFF",
+    justify='center',
+    font=("Motiva Sans Medium", 20 * -1)
+)
+
+canvas.create_text(
+    240.0,
+    456.0,
+    anchor="nw",
+    text="Je SteamID kan je vinden op verschillende websites.\n Om deze functie te kunnen gebruiken heb je de link van je profiel nodig \nDeze kun je aanmaken in je steam instellingen.",
+    fill="#CACACA",
+    justify='center',
+    font=("Motiva Sans Light", 20 * -1)
+)
+
+
+
+window.resizable(False, False)
+window.mainloop()
